requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
flask>=3.0.0
feedparser>=6.0.0
